requirements.txt
accelerate==1.1.1
bitsandbytes==0.44.1
peft==0.13.2
transformers==4.48.3
huggingface_hub==0.26.2
datasets==3.1.0
git-lfs==1.6
wandb==0.18.7
sqlglot==25.32.0
func_timeout==4.3.5
langchain_google_vertexai==2.0.13
python_dotenv==1.0.1
sqlvalidator==0.0.20
vllm==0.7.2
pillow==11.1.0
git+https://github.com/huggingface/trl.git@e95f9fb74a3c3647b86f251b7e230ec51c64b72b#egg=trl\